+++ tests/test_fnmatch.py
@@ -2,6 +2,8 @@
 """Tests for `fnmatch`."""
 import unittest
 import re
+import sys
+import os
 import pytest
 import wcmatch.fnmatch as fnmatch
 from unittest import mock
@@ -479,6 +481,253 @@ class TestFnMatchTranslate(unittest.TestCase):
 
         self.assertTrue(len(fnmatch.translate('!test', flags=fnmatch.N | fnmatch.A)[0]) == 1)
         self.assertTrue(len(fnmatch.translate(b'!test', flags=fnmatch.N | fnmatch.A)[0]) == 1)
+
+
+class TestIsMagic(unittest.TestCase):
+    """Test "is magic" logic."""
+
+    def test_default(self):
+        """Test default magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*"))
+        self.assertTrue(fnmatch.is_magic("test["))
+        self.assertTrue(fnmatch.is_magic("test]"))
+        self.assertTrue(fnmatch.is_magic("test?"))
+        self.assertTrue(fnmatch.is_magic("test\\"))
+
+        self.assertFalse(fnmatch.is_magic("test~!()-/|{}"))
+
+    def test_extmatch(self):
+        """Test extended match magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test[", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test]", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test?", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test\\", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test(", flags=fnmatch.EXTMATCH))
+        self.assertTrue(fnmatch.is_magic("test)", flags=fnmatch.EXTMATCH))
+
+        self.assertFalse(fnmatch.is_magic("test~!-/|{}", flags=fnmatch.EXTMATCH))
+
+    def test_negate(self):
+        """Test negate magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=fnmatch.NEGATE))
+        self.assertTrue(fnmatch.is_magic("test[", flags=fnmatch.NEGATE))
+        self.assertTrue(fnmatch.is_magic("test]", flags=fnmatch.NEGATE))
+        self.assertTrue(fnmatch.is_magic("test?", flags=fnmatch.NEGATE))
+        self.assertTrue(fnmatch.is_magic("test\\", flags=fnmatch.NEGATE))
+        self.assertTrue(fnmatch.is_magic("test!", flags=fnmatch.NEGATE))
+
+        self.assertFalse(fnmatch.is_magic("test~()-/|{}", flags=fnmatch.NEGATE))
+
+    def test_minusnegate(self):
+        """Test minus negate magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+        self.assertTrue(fnmatch.is_magic("test[", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+        self.assertTrue(fnmatch.is_magic("test]", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+        self.assertTrue(fnmatch.is_magic("test?", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+        self.assertTrue(fnmatch.is_magic("test\\", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+        self.assertTrue(fnmatch.is_magic("test-", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+
+        self.assertFalse(fnmatch.is_magic("test~()!/|{}", flags=fnmatch.NEGATE | fnmatch.MINUSNEGATE))
+
+    def test_brace(self):
+        """Test brace magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test[", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test]", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test?", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test\\", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test{", flags=fnmatch.BRACE))
+        self.assertTrue(fnmatch.is_magic("test}", flags=fnmatch.BRACE))
+
+        self.assertFalse(fnmatch.is_magic("test~!-/|", flags=fnmatch.BRACE))
+
+    def test_split(self):
+        """Test split magic."""
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=fnmatch.SPLIT))
+        self.assertTrue(fnmatch.is_magic("test[", flags=fnmatch.SPLIT))
+        self.assertTrue(fnmatch.is_magic("test]", flags=fnmatch.SPLIT))
+        self.assertTrue(fnmatch.is_magic("test?", flags=fnmatch.SPLIT))
+        self.assertTrue(fnmatch.is_magic("test\\", flags=fnmatch.SPLIT))
+        self.assertTrue(fnmatch.is_magic("test|", flags=fnmatch.SPLIT))
+
+        self.assertFalse(fnmatch.is_magic("test~()-!/", flags=fnmatch.SPLIT))
+
+    def test_all(self):
+        """Test tilde magic."""
+
+        flags = (
+            fnmatch.EXTMATCH |
+            fnmatch.NEGATE |
+            fnmatch.BRACE |
+            fnmatch.SPLIT
+        )
+
+        self.assertTrue(fnmatch.is_magic("test*", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test[", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test]", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test?", flags=flags))
+        self.assertTrue(fnmatch.is_magic(r"te\\st", flags=flags))
+        self.assertTrue(fnmatch.is_magic(r"te\st", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test!", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test|", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test(", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test)", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test{", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test}", flags=flags))
+        self.assertTrue(fnmatch.is_magic("test-", flags=flags | fnmatch.MINUSNEGATE))
+
+        self.assertFalse(fnmatch.is_magic("test-~", flags=flags))
+        self.assertFalse(fnmatch.is_magic("test!~", flags=flags | fnmatch.MINUSNEGATE))
+
+    def test_all_bytes(self):
+        """Test tilde magic."""
+
+        flags = (
+            fnmatch.EXTMATCH |
+            fnmatch.NEGATE |
+            fnmatch.BRACE |
+            fnmatch.SPLIT
+        )
+
+        self.assertTrue(fnmatch.is_magic(b"test*", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test[", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test]", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test?", flags=flags))
+        self.assertTrue(fnmatch.is_magic(rb"te\\st", flags=flags))
+        self.assertTrue(fnmatch.is_magic(rb"te\st", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test!", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test|", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test(", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test)", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test{", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test}", flags=flags))
+        self.assertTrue(fnmatch.is_magic(b"test-", flags=flags | fnmatch.MINUSNEGATE))
+
+        self.assertFalse(fnmatch.is_magic(b"test-~", flags=flags))
+        self.assertFalse(fnmatch.is_magic(b"test!~", flags=flags | fnmatch.MINUSNEGATE))
+
+
+class TestFnMatchEscapes(unittest.TestCase):
+    """Test escaping."""
+
+    def check_escape(self, arg, expected, raw=False, unix=None, raw_chars=True):
+        """Verify escapes."""
+
+        flags = 0
+        if unix is False:
+            flags = fnmatch.FORCEWIN
+        elif unix is True:
+            flags = fnmatch.FORCEUNIX
+
+        if raw:
+            self.assertEqual(fnmatch.raw_escape(arg, raw_chars=raw_chars), expected)
+            self.assertEqual(fnmatch.raw_escape(os.fsencode(arg), raw_chars=raw_chars), os.fsencode(expected))
+            file = (util.norm_pattern(arg, False, True) if raw_chars else arg).replace('\\\\', '\\')
+            self.assertTrue(
+                fnmatch.fnmatch(
+                    file,
+                    fnmatch.raw_escape(arg, raw_chars=raw_chars),
+                    flags=flags
+                )
+            )
+        else:
+            self.assertEqual(fnmatch.escape(arg), expected)
+            self.assertEqual(fnmatch.escape(os.fsencode(arg)), os.fsencode(expected))
+            self.assertTrue(
+                fnmatch.fnmatch(
+                    arg,
+                    fnmatch.escape(arg),
+                    flags=flags
+                )
+            )
+
+    def test_escape(self):
+        """Test path escapes."""
+
+        check = self.check_escape
+        check('abc', 'abc')
+        check('[', r'\[')
+        check('?', r'\?')
+        check('*', r'\*')
+        check('[[_/*?*/_]]', r'\[\[_/\*\?\*/_\]\]')
+        check('/[[_/*?*/_]]/', r'/\[\[_/\*\?\*/_\]\]/')
+
+    def test_raw_escape(self):
+        """Test path escapes."""
+
+        check = self.check_escape
+        check(r'abc', 'abc', raw=True)
+        check(r'[', r'\[', raw=True)
+        check(r'?', r'\?', raw=True)
+        check(r'*', r'\*', raw=True)
+        check(r'[[_/*?*/_]]', r'\[\[_/\*\?\*/_\]\]', raw=True)
+        check(r'/[[_/*?*/_]]/', r'/\[\[_/\*\?\*/_\]\]/', raw=True)
+        check(r'\x3f', r'\?', raw=True)
+        # `fnmatch` doesn't care about drives
+        check(r'\\\\[^what]\\name\\temp', r'\\\\\[^what\]\\name\\temp', raw=True, unix=False)
+        check('//[^what]/name/temp', r'//\[^what\]/name/temp', raw=True, unix=False)
+
+    def test_raw_escape_no_raw_chars(self):
+        """Test path escapes with no raw character translations."""
+
+        check = self.check_escape
+        check(r'abc', 'abc', raw=True, raw_chars=False)
+        check(r'[', r'\[', raw=True, raw_chars=False)
+        check(r'?', r'\?', raw=True, raw_chars=False)
+        check(r'*', r'\*', raw=True, raw_chars=False)
+        check(r'[[_/*?*/_]]', r'\[\[_/\*\?\*/_\]\]', raw=True, raw_chars=False)
+        check(r'/[[_/*?*/_]]/', r'/\[\[_/\*\?\*/_\]\]/', raw=True, raw_chars=False)
+        check(r'\x3f', r'\\x3f', raw=True, raw_chars=False)
+        # `fnmatch` doesn't care about drives
+        check(r'\\\\[^what]\\name\\temp', r'\\\\\[^what\]\\name\\temp', raw=True, raw_chars=False, unix=False)
+        check('//[^what]/name/temp', r'//\[^what\]/name/temp', raw=True, raw_chars=False, unix=False)
+
+    @unittest.skipUnless(sys.platform.startswith('win'), "Windows specific test")
+    def test_escape_windows(self):
+        """Test windows escapes."""
+
+        check = self.check_escape
+        # `fnmatch` doesn't care about drives
+        check('a:\\?', r'a:\\\?')
+        check('b:\\*', r'b:\\\*')
+        check('\\\\?\\c:\\?', r'\\\\\?\\c:\\\?')
+        check('\\\\*\\*\\*', r'\\\\\*\\\*\\\*')
+        check('//?/c:/?', r'//\?/c:/\?')
+        check('//*/*/*', r'//\*/\*/\*')
+        check('//[^what]/name/temp', r'//\[^what\]/name/temp')
+
+    def test_escape_forced_windows(self):
+        """Test forced windows escapes."""
+
+        check = self.check_escape
+        # `fnmatch` doesn't care about drives
+        check('a:\\?', r'a:\\\?', unix=False)
+        check('b:\\*', r'b:\\\*', unix=False)
+        check('\\\\?\\c:\\?', r'\\\\\?\\c:\\\?', unix=True)
+        check('\\\\*\\*\\*', r'\\\\\*\\\*\\\*', unix=True)
+        check('//?/c:/?', r'//\?/c:/\?', unix=True)
+        check('//*/*/*', r'//\*/\*/\*', unix=True)
+        check('//[^what]/name/temp', r'//\[^what\]/name/temp', unix=True)
+
+    def test_escape_forced_unix(self):
+        """Test forced windows Unix."""
+
+        check = self.check_escape
+        # `fnmatch` doesn't care about drives
+        check('a:\\?', r'a:\\\?', unix=True)
+        check('b:\\*', r'b:\\\*', unix=True)
+        check('\\\\?\\c:\\?', r'\\\\\?\\c:\\\?', unix=True)
+        check('\\\\*\\*\\*', r'\\\\\*\\\*\\\*', unix=True)
+        check('//?/c:/?', r'//\?/c:/\?', unix=True)
+        check('//*/*/*', r'//\*/\*/\*', unix=True)
+        check('//[^what]/name/temp', r'//\[^what\]/name/temp', unix=True)
 
 
 class TestExpansionLimit(unittest.TestCase):